+++ csas-20230120/core/Models/DeviceGroupDynamicRule.cs
@@ -1,0 +1,26 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Csas20230120.Models
+{
+    public class DeviceGroupDynamicRule : TeaModel {
+        [NameInMap("Args")]
+        [Validation(Required=false)]
+        public object Args { get; set; }
+
+        [NameInMap("Key")]
+        [Validation(Required=false)]
+        public string Key { get; set; }
+
+        [NameInMap("Operator")]
+        [Validation(Required=false)]
+        public string Operator { get; set; }
+
+    }
+
+}